requirements.txt
torch==1.13.0
adapter-transformers==2.2.0
datasets==2.4.0
numpy==1.23.3
tokenizers==0.10.3
tqdm==4.64.1
scikit-learn==1.1.3
